requirements.txt
streamlit
streamlit-option-menu
tensorflow-cpu
numpy
protobuf
pandas
scikit-learn
matplotlib
keras
scikeras
tensorboard